Move WallInTheWay() to Wall.{cpp,h}
