+++ src/openrct2/peep/GuestPathfinding.cpp
@@ -22,6 +22,7 @@
 #include "../world/Entrance.h"
 #include "../world/Footpath.h"
 #include "../world/Map.h"
+#include "../world/Wall.h"
 #include "../world/tile_element/BannerElement.h"
 #include "../world/tile_element/EntranceElement.h"
 #include "../world/tile_element/PathElement.h"
