+++ src/openrct2/world/Wall.cpp
@@ -114,3 +114,36 @@ uint8_t GetWallSlopeFromEdgeSlope(uint8_t Slope, uint8_t Edge)
 {
     return kLandSlopeToWallSlope[Slope][Edge];
 }
+
+/**
+ * Returns true if the edge of tile x, y specified by direction is occupied by a fence
+ * between heights z0 and z1.
+ *
+ * Note that there may still be a fence on the opposing tile.
+ *
+ *  rct2: 0x006E59DC
+ */
+bool WallInTheWay(const CoordsXYRangedZ& fencePos, int32_t direction)
+{
+    TileElement* tileElement;
+
+    tileElement = MapGetFirstElementAt(fencePos);
+    if (tileElement == nullptr)
+        return false;
+    do
+    {
+        if (tileElement->GetType() != TileElementType::Wall)
+            continue;
+        if (tileElement->IsGhost())
+            continue;
+        if (fencePos.baseZ >= tileElement->GetClearanceZ())
+            continue;
+        if (fencePos.clearanceZ <= tileElement->GetBaseZ())
+            continue;
+        if ((tileElement->GetDirection()) != direction)
+            continue;
+
+        return true;
+    } while (!(tileElement++)->IsLastForTile());
+    return false;
+}
